RavenDB-7070 - introduced LegacyFieldIndexing to support importing legacy indexes - removed TransformResultsFunc from AbstractDocumentQuery (not used) - DocumentQuery and AsyncDocumentQuery implementation for SelectFields is now identical - removed Highlighting fields from IndexQuery - moved Highlight/Spatial/Facet methods to appropriate partial classes to remove clutter
