+++ src/Raven.Client/Documents/Session/AsyncDocumentQuery.cs
@@ -1,16 +1,11 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
-using Raven.Client.Documents.Commands;
-using Raven.Client.Documents.Indexes.Spatial;
 using Raven.Client.Documents.Queries;
-using Raven.Client.Documents.Queries.Facets;
-using Raven.Client.Documents.Queries.Spatial;
 using Raven.Client.Documents.Session.Operations;
 using Raven.Client.Documents.Session.Operations.Lazy;
 using Raven.Client.Documents.Session.Tokens;
@@ -22,7 +17,7 @@ namespace Raven.Client.Documents.Session
     /// <summary>
     /// A query against a Raven index
     /// </summary>
-    public class AsyncDocumentQuery<T> : AbstractDocumentQuery<T, AsyncDocumentQuery<T>>, IAsyncDocumentQuery<T>
+    public partial class AsyncDocumentQuery<T> : AbstractDocumentQuery<T, AsyncDocumentQuery<T>>, IAsyncDocumentQuery<T>
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="AsyncDocumentQuery{T}"/> class.
@@ -32,29 +27,21 @@ namespace Raven.Client.Documents.Session
         {
         }
 
-        /// <summary>
-        /// Includes the specified path in the query, loading the document specified in that path
-        /// </summary>
-        /// <param name="path">The path.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Include(string path)
         {
             Include(path);
             return this;
         }
 
-        /// <summary>
-        /// Includes the specified path in the query, loading the document specified in that path
-        /// </summary>
-        /// <param name="path">The path.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Include(Expression<Func<T, object>> path)
         {
             Include(path);
             return this;
         }
 
-        /// <summary>
-        /// Negate the next operation
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Not
         {
             get
@@ -64,484 +51,286 @@ namespace Raven.Client.Documents.Session
             }
         }
 
-        /// <summary>
-        /// Takes the specified count.
-        /// </summary>
-        /// <param name="count">The count.</param>
-        /// <returns></returns>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Take(int count)
         {
             Take(count);
             return this;
         }
 
-        /// <summary>
-        /// Skips the specified count.
-        /// </summary>
-        /// <param name="count">The count.</param>
-        /// <returns></returns>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Skip(int count)
         {
             Skip(count);
             return this;
         }
 
-        /// <summary>
-        /// Filter the results from the index using the specified where clause.
-        /// </summary>
-        /// <param name="whereClause">The where clause.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereLucene(string fieldName, string whereClause)
         {
             WhereLucene(fieldName, whereClause);
             return this;
         }
 
-        /// <summary>
-        /// 	Matches value
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereEquals(string fieldName, object value, bool exact)
         {
             WhereEquals(fieldName, value, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Matches value
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereEquals<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value, bool exact)
         {
             WhereEquals(GetMemberQueryPath(propertySelector.Body), value, exact);
             return this;
         }
 
-        /// <summary>
-        /// Matches value
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereEquals(WhereParams whereParams)
         {
             WhereEquals(whereParams);
             return this;
         }
 
-        /// <summary>
-        /// 	Not matches value
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereNotEquals(string fieldName, object value, bool exact)
         {
             WhereNotEquals(fieldName, value, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Not matches value
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereNotEquals<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value, bool exact)
         {
             WhereNotEquals(GetMemberQueryPath(propertySelector.Body), value, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Not matches value
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereNotEquals(WhereParams whereParams)
         {
             WhereNotEquals(whereParams);
             return this;
         }
 
-        /// <summary>
-        /// Check that the field has one of the specified value
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereIn(string fieldName, IEnumerable<object> values, bool exact)
         {
             WhereIn(fieldName, values, exact);
             return this;
         }
 
-        /// <summary>
-        /// Check that the field has one of the specified value
-        /// </summary>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> WhereIn<TValue>(Expression<Func<T, TValue>> propertySelector, IEnumerable<TValue> values, bool exact = false)
         {
             WhereIn(GetMemberQueryPath(propertySelector.Body), values.Cast<object>(), exact);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields which starts with the specified value.
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="value">The value.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereStartsWith(string fieldName, object value)
         {
             WhereStartsWith(fieldName, value);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields which starts with the specified value.
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "value">The value.</param>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> WhereStartsWith<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value)
         {
             WhereStartsWith(GetMemberQueryPath(propertySelector.Body), value);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields which ends with the specified value.
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="value">The value.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereEndsWith(string fieldName, object value)
         {
             WhereEndsWith(fieldName, value);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields which ends with the specified value.
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "value">The value.</param>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> WhereEndsWith<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value)
         {
             WhereEndsWith(GetMemberQueryPath(propertySelector.Body), value);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields where the value is between the specified start and end, exclusive
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="start">The start.</param>
-        /// <param name="end">The end.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereBetween(string fieldName, object start, object end, bool exact)
         {
             WhereBetween(fieldName, start, end, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields where the value is between the specified start and end, exclusive
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "start">The start.</param>
-        /// <param name = "end">The end.</param>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> WhereBetween<TValue>(Expression<Func<T, TValue>> propertySelector, TValue start, TValue end, bool exact = false)
         {
             WhereBetween(GetMemberQueryPath(propertySelector.Body), start, end, exact);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields where the value is greater than the specified value
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="value">The value.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereGreaterThan(string fieldName, object value, bool exact)
         {
             WhereGreaterThan(fieldName, value, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields where the value is greater than the specified value
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "value">The value.</param>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> WhereGreaterThan<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value, bool exact = false)
         {
             WhereGreaterThan(GetMemberQueryPath(propertySelector.Body), value, exact);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields where the value is greater than or equal to the specified value
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="value">The value.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereGreaterThanOrEqual(string fieldName, object value, bool exact)
         {
             WhereGreaterThanOrEqual(fieldName, value, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields where the value is greater than or equal to the specified value
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "value">The value.</param>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> WhereGreaterThanOrEqual<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value, bool exact = false)
         {
             WhereGreaterThanOrEqual(GetMemberQueryPath(propertySelector.Body), value, exact);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields where the value is less than the specified value
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="value">The value.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereLessThan(string fieldName, object value, bool exact)
         {
             WhereLessThan(fieldName, value, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields where the value is less than the specified value
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "value">The value.</param>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> WhereLessThan<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value, bool exact = false)
         {
             WhereLessThan(GetMemberQueryPath(propertySelector.Body), value, exact);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields where the value is less than or equal to the specified value
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="value">The value.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereLessThanOrEqual(string fieldName, object value, bool exact)
         {
             WhereLessThanOrEqual(fieldName, value, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields where the value is less than or equal to the specified value
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "value">The value.</param>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> WhereLessThanOrEqual<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value, bool exact = false)
         {
             WhereLessThanOrEqual(GetMemberQueryPath(propertySelector.Body), value, exact);
             return this;
         }
 
-        /// <summary>
-        ///     Check if the given field exists
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> WhereExists<TValue>(Expression<Func<T, TValue>> propertySelector)
         {
             WhereExists(GetMemberQueryPath(propertySelector.Body));
             return this;
         }
 
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WhereExists(string fieldName)
         {
             WhereExists(fieldName);
             return this;
         }
 
-        /// <summary>
-        /// Add an AND to the query
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.AndAlso()
         {
             AndAlso();
             return this;
         }
 
-        /// <summary>
-        /// Add an OR to the query
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrElse()
         {
             OrElse();
             return this;
         }
 
-        /// <summary>
-        /// Specifies a boost weight to the last where clause.
-        /// The higher the boost factor, the more relevant the term will be.
-        /// </summary>
-        /// <param name="boost">boosting factor where 1.0 is default, less than 1.0 is lower weight, greater than 1.0 is higher weight</param>
-        /// <returns></returns>
-        /// <remarks>
-        /// http://lucene.apache.org/java/2_4_0/queryparsersyntax.html#Boosting%20a%20Term
-        /// </remarks>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Boost(decimal boost)
         {
             Boost(boost);
             return this;
         }
 
-        /// <summary>
-        /// Specifies a fuzziness factor to the single word term in the last where clause
-        /// </summary>
-        /// <param name="fuzzy">0.0 to 1.0 where 1.0 means closer match</param>
-        /// <returns></returns>
-        /// <remarks>
-        /// http://lucene.apache.org/java/2_4_0/queryparsersyntax.html#Fuzzy%20Searches
-        /// </remarks>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Fuzzy(decimal fuzzy)
         {
             Fuzzy(fuzzy);
             return this;
         }
 
-        /// <summary>
-        /// Specifies a proximity distance for the phrase in the last where clause
-        /// </summary>
-        /// <param name="proximity">number of words within</param>
-        /// <returns></returns>
-        /// <remarks>
-        /// http://lucene.apache.org/java/2_4_0/queryparsersyntax.html#Proximity%20Searches
-        /// </remarks>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Proximity(int proximity)
         {
             Proximity(proximity);
             return this;
         }
 
-        /// <summary>
-        /// Order the search results randomly
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.RandomOrdering()
         {
             RandomOrdering();
             return this;
         }
 
-        /// <summary>
-        /// Order the search results randomly using the specified seed
-        /// this is useful if you want to have repeatable random queries
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.RandomOrdering(string seed)
         {
             RandomOrdering(seed);
             return this;
         }
 
-        /// <summary>
-        /// Order the search results randomly
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.CustomSortUsing(string typeName, bool descending)
         {
             CustomSortUsing(typeName, descending);
             return this;
         }
 
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WithinRadiusOf<TValue>(Expression<Func<T, TValue>> propertySelector, double radius, double latitude, double longitude, SpatialUnits? radiusUnits, double distanceErrorPct)
-        {
-            WithinRadiusOf(propertySelector.ToPropertyPath(), radius, latitude, longitude, radiusUnits, distanceErrorPct);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WithinRadiusOf(string fieldName, double radius, double latitude, double longitude, SpatialUnits? radiusUnits, double distanceErrorPct)
-        {
-            WithinRadiusOf(fieldName, radius, latitude, longitude, radiusUnits, distanceErrorPct);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.RelatesToShape<TValue>(Expression<Func<T, TValue>> propertySelector, string shapeWKT, SpatialRelation relation, double distanceErrorPct)
-        {
-            Spatial(propertySelector.ToPropertyPath(), shapeWKT, relation, distanceErrorPct);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.RelatesToShape(string fieldName, string shapeWKT, SpatialRelation relation, double distanceErrorPct)
-        {
-            Spatial(fieldName, shapeWKT, relation, distanceErrorPct);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrderByDistance<TValue>(Expression<Func<T, TValue>> propertySelector, double latitude, double longitude)
-        {
-            OrderByDistance(propertySelector.ToPropertyPath(), latitude, longitude);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrderByDistance(string fieldName, double latitude, double longitude)
-        {
-            OrderByDistance(fieldName, latitude, longitude);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrderByDistance<TValue>(Expression<Func<T, TValue>> propertySelector, string shapeWkt)
-        {
-            OrderByDistance(propertySelector.ToPropertyPath(), shapeWkt);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrderByDistance(string fieldName, string shapeWkt)
-        {
-            OrderByDistance(fieldName, shapeWkt);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrderByDistanceDescending<TValue>(Expression<Func<T, TValue>> propertySelector, double latitude, double longitude)
-        {
-            OrderByDistanceDescending(propertySelector.ToPropertyPath(), latitude, longitude);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrderByDistanceDescending(string fieldName, double latitude, double longitude)
-        {
-            OrderByDistanceDescending(fieldName, latitude, longitude);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrderByDistanceDescending<TValue>(Expression<Func<T, TValue>> propertySelector, string shapeWkt)
-        {
-            OrderByDistanceDescending(propertySelector.ToPropertyPath(), shapeWkt);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrderByDistanceDescending(string fieldName, string shapeWkt)
-        {
-            OrderByDistanceDescending(fieldName, shapeWkt);
-            return this;
-        }
-
+        /// <inheritdoc />
         public IAsyncDocumentQuery<TResult> OfType<TResult>()
         {
             return CreateDocumentQueryInternal<TResult>();
         }
 
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IAsyncDocumentQuery<T>.RawQuery(string query)
         {
             RawQuery(query);
             return this;
         }
 
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IAsyncDocumentQuery<T>.AddParameter(string name, object value)
         {
             AddParameter(name, value);
             return this;
         }
 
+        /// <inheritdoc />
         IAsyncGroupByDocumentQuery<T> IAsyncDocumentQuery<T>.GroupBy(string fieldName, params string[] fieldNames)
         {
             GroupBy(fieldName, fieldNames);
             return new AsyncGroupByDocumentQuery<T>(this);
         }
 
-        /// <summary>
-        /// Order the results by the specified fields
-        /// The fields are the names of the fields to sort, defaulting to sorting by ascending.
-        /// You can prefix a field name with '-' to indicate sorting by descending or '+' to sort by ascending
-        /// </summary>
-        /// <param name="fields">The fields.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrderBy(string field, OrderingType ordering)
         {
             OrderBy(field, ordering);
             return this;
         }
 
-        /// <summary>
-        ///   Order the results by the specified fields
-        ///   The fields are the names of the fields to sort, defaulting to sorting by ascending.
-        ///   You can prefix a field name with '-' to indicate sorting by descending or '+' to sort by ascending
-        /// </summary>
-        /// <param name = "propertySelectors">Property selectors for the fields.</param>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> OrderBy<TValue>(params Expression<Func<T, TValue>>[] propertySelectors)
         {
             foreach (var item in propertySelectors)
@@ -552,24 +341,14 @@ namespace Raven.Client.Documents.Session
             return this;
         }
 
-        /// <summary>
-        /// Order the results by the specified fields
-        /// The fields are the names of the fields to sort, defaulting to sorting by descending.
-        /// You can prefix a field name with '-' to indicate sorting by descending or '+' to sort by ascending
-        /// </summary>
-        /// <param name="fields">The fields.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrderByDescending(string field, OrderingType ordering)
         {
             OrderByDescending(field, ordering);
             return this;
         }
 
-        /// <summary>
-        ///   Order the results by the specified fields
-        ///   The fields are the names of the fields to sort, defaulting to sorting by descending.
-        ///   You can prefix a field name with '-' to indicate sorting by descending or '+' to sort by ascending
-        /// </summary>
-        /// <param name = "propertySelectors">Property selectors for the fields.</param>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> OrderByDescending<TValue>(params Expression<Func<T, TValue>>[] propertySelectors)
         {
             foreach (var item in propertySelectors)
@@ -580,183 +359,78 @@ namespace Raven.Client.Documents.Session
             return this;
         }
 
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Highlight(
-            string fieldName, int fragmentLength, int fragmentCount, string fragmentsField)
-        {
-            Highlight(fieldName, fragmentLength, fragmentCount, fragmentsField);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Highlight(string fieldName, int fragmentLength, int fragmentCount,
-            out FieldHighlightings fieldHighlightings)
-        {
-            Highlight(fieldName, fragmentLength, fragmentCount, out fieldHighlightings);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Highlight(string fieldName, string fieldKeyName, int fragmentLength, int fragmentCount,
-            out FieldHighlightings fieldHighlightings)
-        {
-            Highlight(fieldName, fieldKeyName, fragmentLength, fragmentCount, out fieldHighlightings);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Highlight<TValue>(
-            Expression<Func<T, TValue>> propertySelector,
-            int fragmentLength,
-            int fragmentCount,
-            Expression<Func<T, IEnumerable>> fragmentsPropertySelector)
-        {
-            var fieldName = GetMemberQueryPath(propertySelector);
-            var fragmentsField = GetMemberQueryPath(fragmentsPropertySelector);
-            Highlight(fieldName, fragmentLength, fragmentCount, fragmentsField);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Highlight<TValue>(
-            Expression<Func<T, TValue>> propertySelector, int fragmentLength, int fragmentCount,
-            out FieldHighlightings fieldHighlightings)
-        {
-            Highlight(GetMemberQueryPath(propertySelector), fragmentLength, fragmentCount, out fieldHighlightings);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Highlight<TValue>(
-            Expression<Func<T, TValue>> propertySelector, Expression<Func<T, TValue>> keyPropertySelector, int fragmentLength, int fragmentCount,
-            out FieldHighlightings fieldHighlightings)
-        {
-            Highlight(GetMemberQueryPath(propertySelector), GetMemberQueryPath(keyPropertySelector), fragmentLength, fragmentCount, out fieldHighlightings);
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.SetHighlighterTags(string preTag, string postTag)
-        {
-            SetHighlighterTags(new[] { preTag }, new[] { postTag });
-            return this;
-        }
-
-        IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.SetHighlighterTags(string[] preTags, string[] postTags)
-        {
-            SetHighlighterTags(preTags, postTags);
-            return this;
-        }
-
-        /// <summary>
-        /// Instructs the query to wait for non stale results as of now.
-        /// </summary>
-        /// <returns></returns>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WaitForNonStaleResultsAsOfNow()
         {
             WaitForNonStaleResultsAsOfNow();
             return this;
         }
 
-        /// <summary>
-        /// Instructs the query to wait for non stale results as of now for the specified timeout.
-        /// </summary>
-        /// <param name="waitTimeout">The wait timeout.</param>
-        /// <returns></returns>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WaitForNonStaleResultsAsOfNow(TimeSpan waitTimeout)
         {
             WaitForNonStaleResultsAsOfNow(waitTimeout);
             return this;
         }
 
-        /// <summary>
-        /// Instructs the query to wait for non stale results as of the cutoff etag.
-        /// </summary>
-        /// <param name="cutOffEtag">The cut off etag.</param>
-        /// <returns></returns>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WaitForNonStaleResultsAsOf(long cutOffEtag)
         {
             WaitForNonStaleResultsAsOf(cutOffEtag);
             return this;
         }
 
-        /// <summary>
-        /// Instructs the query to wait for non stale results as of the cutoff etag for the specified timeout.
-        /// </summary>
-        /// <param name="cutOffEtag">The cut off etag.</param>
-        /// <param name="waitTimeout">The wait timeout.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WaitForNonStaleResultsAsOf(long cutOffEtag, TimeSpan waitTimeout)
         {
             WaitForNonStaleResultsAsOf(cutOffEtag, waitTimeout);
             return this;
         }
 
-        /// <summary>
-        /// EXPERT ONLY: Instructs the query to wait for non stale results.
-        /// This shouldn't be used outside of unit tests unless you are well aware of the implications
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WaitForNonStaleResults()
         {
             WaitForNonStaleResults();
             return this;
         }
 
-        /// <summary>
-        /// Allows you to modify the index query before it is sent to the server
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.BeforeQueryExecution(Action<IndexQuery> beforeQueryExecution)
         {
             BeforeQueryExecution(beforeQueryExecution);
             return this;
         }
 
-        /// <summary>
-        /// EXPERT ONLY: Instructs the query to wait for non stale results for the specified wait timeout.
-        /// This shouldn't be used outside of unit tests unless you are well aware of the implications
-        /// </summary>
-        /// <param name="waitTimeout">The wait timeout.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.WaitForNonStaleResults(TimeSpan waitTimeout)
         {
             WaitForNonStaleResults(waitTimeout);
             return this;
         }
 
-        /// <summary>
-        /// Selects all the projection fields directly from the index
-        /// </summary>
-        /// <typeparam name="TProjection">The type of the projection.</typeparam>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<TProjection> SelectFields<TProjection>()
         {
-            return SelectFields<TProjection>(ReflectionUtil.GetPropertiesAndFieldsFor<TProjection>(BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public).Select(x => x.Name).ToArray());
+            var propertyInfos = ReflectionUtil.GetPropertiesAndFieldsFor<TProjection>(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance).ToList();
+            var projections = propertyInfos.Select(x => x.Name).ToArray();
+            var identityProperty = Conventions.GetIdentityProperty(typeof(TProjection));
+            var fields = propertyInfos.Select(p => p == identityProperty ? Constants.Documents.Indexing.Fields.DocumentIdFieldName : p.Name).ToArray();
+            return SelectFields<TProjection>(fields, projections);
         }
 
-        /// <summary>
-        /// Selects the specified fields directly from the index
-        /// </summary>
-        /// <typeparam name="TProjection">The type of the projection.</typeparam>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<TProjection> SelectFields<TProjection>(params string[] fields)
         {
             return SelectFields<TProjection>(fields, fields);
         }
 
-        /// <summary>
-        /// Selects the specified fields directly from the index
-        /// </summary>
-        /// <typeparam name="TProjection">The type of the projection.</typeparam>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<TProjection> SelectFields<TProjection>(string[] fields, string[] projections)
         {
             return CreateDocumentQueryInternal<TProjection>(fields.Length > 0 ? FieldsToFetchToken.Create(fields, projections) : null);
         }
 
-        public IAsyncDocumentQuery<T> Spatial(Expression<Func<T, object>> path, Func<SpatialCriteriaFactory, SpatialCriteria> clause)
-        {
-            return Spatial(path.ToPropertyPath(), clause);
-        }
-
-        public IAsyncDocumentQuery<T> Spatial(string fieldName, Func<SpatialCriteriaFactory, SpatialCriteria> clause)
-        {
-            var criteria = clause(SpatialCriteriaFactory.Instance);
-            Spatial(fieldName, criteria);
-            return this;
-        }
-
-        /// <summary>
-        /// Register the query as a lazy-count query in the session and return a lazy
-        /// instance that will evaluate the query only when needed
-        /// </summary>
+        /// <inheritdoc />
         public Lazy<Task<int>> CountLazilyAsync(CancellationToken token = default(CancellationToken))
         {
             if (QueryOperation == null)
@@ -770,12 +444,7 @@ namespace Raven.Client.Documents.Session
             return ((AsyncDocumentSession)TheSession).AddLazyCountOperation(lazyQueryOperation, token);
         }
 
-        /// <summary>
-        /// Adds an ordering for a specific field to the query
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="descending">if set to <c>true</c> [descending].</param>
-        /// <param name="ordering">Ordering type.</param>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.AddOrder(string fieldName, bool descending, OrderingType ordering)
         {
             if (descending)
@@ -786,30 +455,21 @@ namespace Raven.Client.Documents.Session
             return this;
         }
 
-        /// <summary>
-        /// Adds an ordering by score for a specific field to the query
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrderByScore()
         {
             OrderByScore();
             return this;
         }
 
-        /// <summary>
-        /// Adds an ordering by score descending for a specific field to the query
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OrderByScoreDescending()
         {
             OrderByScoreDescending();
             return this;
         }
 
-        /// <summary>
-        ///   Adds an ordering for a specific field to the query
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "descending">if set to <c>true</c> [descending].</param>
-        /// <param name = "ordering">Ordering type.</param>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> AddOrder<TValue>(Expression<Func<T, TValue>> propertySelector, bool descending, OrderingType ordering)
         {
             var fieldName = GetMemberQueryPath(propertySelector.Body);
@@ -821,181 +481,119 @@ namespace Raven.Client.Documents.Session
             return this;
         }
 
-        /// <summary>
-        /// Simplified method for opening a new clause within the query
-        /// </summary>
-        /// <returns></returns>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.OpenSubclause()
         {
             OpenSubclause();
             return this;
         }
 
-        /// <summary>
-        /// Perform a search for documents which fields that match the searchTerms.
-        /// If there is more than a single term, each of them will be checked independently.
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Search(string fieldName, string searchTerms, SearchOperator @operator)
         {
             Search(fieldName, searchTerms, @operator);
             return this;
         }
 
-        /// <summary>
-        /// Perform a search for documents which fields that match the searchTerms.
-        /// If there is more than a single term, each of them will be checked independently.
-        /// </summary>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> Search<TValue>(Expression<Func<T, TValue>> propertySelector, string searchTerms, SearchOperator @operator)
         {
             Search(GetMemberQueryPath(propertySelector.Body), searchTerms, @operator);
             return this;
         }
 
-        /// <summary>
-        /// Simplified method for closing a clause within the query
-        /// </summary>
-        /// <returns></returns>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.CloseSubclause()
         {
             CloseSubclause();
             return this;
         }
 
-        /// <summary>
-        /// Partition the query so we can intersect different parts of the query
-        /// across different index entries.
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Intersect()
         {
             Intersect();
             return this;
         }
 
-        /// <summary>
-        /// Performs a query matching ANY of the provided values against the given field (OR)
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.ContainsAny(string fieldName, IEnumerable<object> values)
         {
             ContainsAny(fieldName, values);
             return this;
         }
 
-        /// <summary>
-        /// Performs a query matching ANY of the provided values against the given field (OR)
-        /// </summary>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> ContainsAny<TValue>(Expression<Func<T, TValue>> propertySelector, IEnumerable<TValue> values)
         {
             ContainsAny(GetMemberQueryPath(propertySelector.Body), values.Cast<object>());
             return this;
         }
 
-        /// <summary>
-        /// Performs a query matching ALL of the provided values against the given field (AND)
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.ContainsAll(string fieldName, IEnumerable<object> values)
         {
             ContainsAll(fieldName, values);
             return this;
         }
 
-        /// <summary>
-        /// Performs a query matching ALL of the provided values against the given field (AND)
-        /// </summary>
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> ContainsAll<TValue>(Expression<Func<T, TValue>> propertySelector, IEnumerable<TValue> values)
         {
             ContainsAll(GetMemberQueryPath(propertySelector.Body), values.Cast<object>());
             return this;
         }
 
-        /// <summary>
-        /// Provide statistics about the query, such as total count of matching records
-        /// </summary>
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Statistics(out QueryStatistics stats)
         {
             Statistics(out stats);
             return this;
         }
 
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.UsingDefaultOperator(QueryOperator queryOperator)
         {
             UsingDefaultOperator(queryOperator);
             return this;
         }
 
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.NoTracking()
         {
             NoTracking();
             return this;
         }
 
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.NoCaching()
         {
             NoCaching();
             return this;
         }
 
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.ShowTimings()
         {
             ShowTimings();
             return this;
         }
 
+        /// <inheritdoc />
         IAsyncDocumentQuery<T> IDocumentQueryBase<T, IAsyncDocumentQuery<T>>.Distinct()
         {
             Distinct();
             return this;
         }
 
+        /// <inheritdoc />
         public IAsyncDocumentQuery<T> ExplainScores()
         {
             ShouldExplainScores = true;
             return this;
-
         }
 
-        public async Task<FacetedQueryResult> GetFacetsAsync(string facetSetupDoc, int facetStart, int? facetPageSize, CancellationToken token = default(CancellationToken))
-        {
-            var q = GetIndexQuery();
-            var query = FacetQuery.Create(q, facetSetupDoc, null, facetStart, facetPageSize, Conventions);
-
-            var command = new GetFacetsCommand(Conventions, TheSession.Context, query);
-            await TheSession.RequestExecutor.ExecuteAsync(command, TheSession.Context, token).ConfigureAwait(false);
-
-            return command.Result;
-        }
-
-        public async Task<FacetedQueryResult> GetFacetsAsync(List<Facet> facets, int facetStart, int? facetPageSize, CancellationToken token = default(CancellationToken))
-        {
-            var q = GetIndexQuery();
-            var query = FacetQuery.Create(q, null, facets, facetStart, facetPageSize, Conventions);
-
-            var command = new GetFacetsCommand(Conventions, TheSession.Context, query);
-            await TheSession.RequestExecutor.ExecuteAsync(command, TheSession.Context, token).ConfigureAwait(false);
-
-            return command.Result;
-        }
-
-        public Lazy<Task<FacetedQueryResult>> GetFacetsLazyAsync(string facetSetupDoc, int facetStart, int? facetPageSize, CancellationToken token = default(CancellationToken))
-        {
-            var q = GetIndexQuery();
-            var query = FacetQuery.Create(q, facetSetupDoc, null, facetStart, facetPageSize, Conventions);
-
-            var lazyFacetsOperation = new LazyFacetsOperation(Conventions, query);
-            return ((AsyncDocumentSession)TheSession).AddLazyOperation<FacetedQueryResult>(lazyFacetsOperation, null, token);
-        }
-
-        public Lazy<Task<FacetedQueryResult>> GetFacetsLazyAsync(List<Facet> facets, int facetStart, int? facetPageSize, CancellationToken token = default(CancellationToken))
-        {
-            var q = GetIndexQuery();
-            var query = FacetQuery.Create(q, null, facets, facetStart, facetPageSize, Conventions);
-
-            var lazyFacetsOperation = new LazyFacetsOperation(Conventions, query);
-            return ((AsyncDocumentSession)TheSession).AddLazyOperation<FacetedQueryResult>(lazyFacetsOperation, null, token);
-        }
-
-        /// <summary>
-        /// Returns a list of results for a query asynchronously. 
-        /// </summary>
+        /// <inheritdoc />
         public async Task<List<T>> ToListAsync(CancellationToken token = default(CancellationToken))
         {
             await InitAsync(token).ConfigureAwait(false);
@@ -1003,24 +601,28 @@ namespace Raven.Client.Documents.Session
             return tuple.Item2;
         }
 
+        /// <inheritdoc />
         public async Task<T> FirstAsync(CancellationToken token = default(CancellationToken))
         {
             var operation = await ExecuteQueryOperation(1, token).ConfigureAwait(false);
             return operation.First();
         }
 
+        /// <inheritdoc />
         public async Task<T> FirstOrDefaultAsync(CancellationToken token = default(CancellationToken))
         {
             var operation = await ExecuteQueryOperation(1, token).ConfigureAwait(false);
             return operation.FirstOrDefault();
         }
 
+        /// <inheritdoc />
         public async Task<T> SingleAsync(CancellationToken token = default(CancellationToken))
         {
             var operation = await ExecuteQueryOperation(2, token).ConfigureAwait(false);
             return operation.Single();
         }
 
+        /// <inheritdoc />
         public async Task<T> SingleOrDefaultAsync(CancellationToken token = default(CancellationToken))
         {
             var operation = await ExecuteQueryOperation(2, token).ConfigureAwait(false);
@@ -1037,10 +639,7 @@ namespace Raven.Client.Documents.Session
             return QueryOperation.Complete<T>();
         }
 
-        /// <summary>
-        /// Register the query as a lazy query in the session and return a lazy
-        /// instance that will evaluate the query only when needed
-        /// </summary>
+        /// <inheritdoc />
         public Lazy<Task<IEnumerable<T>>> LazilyAsync(Action<IEnumerable<T>> onEval = null)
         {
             if (QueryOperation == null)
@@ -1052,9 +651,7 @@ namespace Raven.Client.Documents.Session
             return ((AsyncDocumentSession)TheSession).AddLazyOperation(lazyQueryOperation, onEval);
         }
 
-        /// <summary>
-        /// Gets the total count of records for this query
-        /// </summary>
+        /// <inheritdoc />
         public async Task<int> CountAsync(CancellationToken token = default(CancellationToken))
         {
             Take(0);
@@ -1068,11 +665,7 @@ namespace Raven.Client.Documents.Session
             return Task.FromResult(Tuple.Create(currentQueryOperation.CurrentQueryResults, list));
         }
 
-        /// <summary>
-        ///   Gets the query result
-        ///   Execute the query the first time that this is called.
-        /// </summary>
-        /// <value>The query result.</value>
+        /// <inheritdoc />
         public async Task<QueryResult> QueryResultAsync(CancellationToken token = default(CancellationToken))
         {
             await InitAsync(token).ConfigureAwait(false);
@@ -1080,7 +673,7 @@ namespace Raven.Client.Documents.Session
             return QueryOperation.CurrentQueryResults.CreateSnapshot();
         }
 
-        protected virtual async Task InitAsync(CancellationToken token)
+        protected async Task InitAsync(CancellationToken token)
         {
             if (QueryOperation != null)
                 return;
@@ -1130,7 +723,6 @@ namespace Raven.Client.Documents.Session
                 TheWaitForNonStaleResults = TheWaitForNonStaleResults,
                 TheWaitForNonStaleResultsAsOfNow = TheWaitForNonStaleResultsAsOfNow,
                 Negate = Negate,
-                TransformResultsFunc = TransformResultsFunc,
                 Includes = new HashSet<string>(Includes),
                 RootTypes = { typeof(T) },
                 BeforeQueryExecutionAction = BeforeQueryExecutionAction,
