+++ src/Raven.Client/Documents/Session/DocumentQuery.Highlight.cs
@@ -1,0 +1,93 @@
+﻿using System;
+using System.Collections;
+using System.Linq.Expressions;
+
+namespace Raven.Client.Documents.Session
+{
+    public partial class DocumentQuery<T>
+    {
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Highlight(
+            string fieldName,
+            int fragmentLength,
+            int fragmentCount,
+            string fragmentsField)
+        {
+            Highlight(fieldName, fragmentLength, fragmentCount, fragmentsField);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Highlight(
+            string fieldName,
+            int fragmentLength,
+            int fragmentCount,
+            out FieldHighlightings highlightings)
+        {
+            Highlight(fieldName, fragmentLength, fragmentCount, out highlightings);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Highlight(
+            string fieldName,
+            string fieldKeyName,
+            int fragmentLength,
+            int fragmentCount,
+            out FieldHighlightings highlightings)
+        {
+            Highlight(fieldName, fieldKeyName, fragmentLength, fragmentCount, out highlightings);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Highlight<TValue>(
+            Expression<Func<T, TValue>> propertySelector,
+            int fragmentLength,
+            int fragmentCount,
+            Expression<Func<T, IEnumerable>> fragmentsPropertySelector)
+        {
+            var fieldName = GetMemberQueryPath(propertySelector);
+            var fragmentsField = GetMemberQueryPath(fragmentsPropertySelector);
+            Highlight(fieldName, fragmentLength, fragmentCount, fragmentsField);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Highlight<TValue>(
+            Expression<Func<T, TValue>> propertySelector,
+            int fragmentLength,
+            int fragmentCount,
+            out FieldHighlightings fieldHighlightings)
+        {
+            Highlight(GetMemberQueryPath(propertySelector), fragmentLength, fragmentCount, out fieldHighlightings);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Highlight<TValue>(
+            Expression<Func<T, TValue>> propertySelector,
+            Expression<Func<T, TValue>> keyPropertySelector,
+            int fragmentLength,
+            int fragmentCount,
+            out FieldHighlightings fieldHighlightings)
+        {
+            Highlight(GetMemberQueryPath(propertySelector), GetMemberQueryPath(keyPropertySelector), fragmentLength, fragmentCount, out fieldHighlightings);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.SetHighlighterTags(string preTag, string postTag)
+        {
+            SetHighlighterTags(new[] { preTag }, new[] { postTag });
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.SetHighlighterTags(string[] preTags, string[] postTags)
+        {
+            SetHighlighterTags(preTags, postTags);
+            return this;
+        }
+    }
+}
