+++ src/Raven.Client/Documents/Session/DocumentQuery.Spatial.cs
@@ -1,0 +1,109 @@
+﻿using System;
+using System.Linq.Expressions;
+using Raven.Client.Documents.Indexes.Spatial;
+using Raven.Client.Documents.Queries.Spatial;
+using Raven.Client.Extensions;
+
+namespace Raven.Client.Documents.Session
+{
+    public partial class DocumentQuery<T>
+    {
+        /// <inheritdoc />
+        public IDocumentQuery<T> Spatial(Expression<Func<T, object>> path, Func<SpatialCriteriaFactory, SpatialCriteria> clause)
+        {
+            return Spatial(path.ToPropertyPath(), clause);
+        }
+
+        /// <inheritdoc />
+        public IDocumentQuery<T> Spatial(string fieldName, Func<SpatialCriteriaFactory, SpatialCriteria> clause)
+        {
+            var criteria = clause(SpatialCriteriaFactory.Instance);
+            Spatial(fieldName, criteria);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WithinRadiusOf<TValue>(Expression<Func<T, TValue>> propertySelector, double radius, double latitude, double longitude, SpatialUnits? radiusUnits, double distanceErrorPct)
+        {
+            WithinRadiusOf(propertySelector.ToPropertyPath(), radius, latitude, longitude, radiusUnits, distanceErrorPct);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WithinRadiusOf(string fieldName, double radius, double latitude, double longitude, SpatialUnits? radiusUnits, double distanceErrorPct)
+        {
+            WithinRadiusOf(fieldName, radius, latitude, longitude, radiusUnits, distanceErrorPct);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.RelatesToShape<TValue>(Expression<Func<T, TValue>> propertySelector, string shapeWKT, SpatialRelation relation, double distanceErrorPct)
+        {
+            Spatial(propertySelector.ToPropertyPath(), shapeWKT, relation, distanceErrorPct);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.RelatesToShape(string fieldName, string shapeWKT, SpatialRelation relation, double distanceErrorPct)
+        {
+            Spatial(fieldName, shapeWKT, relation, distanceErrorPct);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistance<TValue>(Expression<Func<T, TValue>> propertySelector, double latitude, double longitude)
+        {
+            OrderByDistance(propertySelector.ToPropertyPath(), latitude, longitude);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistance(string fieldName, double latitude, double longitude)
+        {
+            OrderByDistance(fieldName, latitude, longitude);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistance<TValue>(Expression<Func<T, TValue>> propertySelector, string shapeWkt)
+        {
+            OrderByDistance(propertySelector.ToPropertyPath(), shapeWkt);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistance(string fieldName, string shapeWkt)
+        {
+            OrderByDistance(fieldName, shapeWkt);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistanceDescending<TValue>(Expression<Func<T, TValue>> propertySelector, double latitude, double longitude)
+        {
+            OrderByDistanceDescending(propertySelector.ToPropertyPath(), latitude, longitude);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistanceDescending(string fieldName, double latitude, double longitude)
+        {
+            OrderByDistanceDescending(fieldName, latitude, longitude);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistanceDescending<TValue>(Expression<Func<T, TValue>> propertySelector, string shapeWkt)
+        {
+            OrderByDistanceDescending(propertySelector.ToPropertyPath(), shapeWkt);
+            return this;
+        }
+
+        /// <inheritdoc />
+        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistanceDescending(string fieldName, string shapeWkt)
+        {
+            OrderByDistanceDescending(fieldName, shapeWkt);
+            return this;
+        }
+    }
+}
