+++ src/Raven.Client/Documents/Session/DocumentQuery.cs
@@ -4,14 +4,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
-using Raven.Client.Documents.Commands;
-using Raven.Client.Documents.Indexes.Spatial;
 using Raven.Client.Documents.Queries;
-using Raven.Client.Documents.Queries.Facets;
-using Raven.Client.Documents.Queries.Spatial;
 using Raven.Client.Documents.Session.Operations.Lazy;
 using Raven.Client.Documents.Session.Tokens;
-using Raven.Client.Extensions;
 using Raven.Client.Util;
 
 namespace Raven.Client.Documents.Session
@@ -19,7 +14,7 @@ namespace Raven.Client.Documents.Session
     /// <summary>
     /// A query against a Raven index
     /// </summary>
-    public class DocumentQuery<T> : AbstractDocumentQuery<T, DocumentQuery<T>>, IDocumentQuery<T>
+    public partial class DocumentQuery<T> : AbstractDocumentQuery<T, DocumentQuery<T>>, IDocumentQuery<T>
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="DocumentQuery{T}"/> class.
@@ -29,78 +24,64 @@ namespace Raven.Client.Documents.Session
         {
         }
 
-        /// <summary>
-        /// Selects the projection fields directly from the index
-        /// </summary>
-        /// <typeparam name="TProjection">The type of the projection.</typeparam>
+        /// <inheritdoc />
         public IDocumentQuery<TProjection> SelectFields<TProjection>()
         {
             var propertyInfos = ReflectionUtil.GetPropertiesAndFieldsFor<TProjection>(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance).ToList();
             var projections = propertyInfos.Select(x => x.Name).ToArray();
             var identityProperty = Conventions.GetIdentityProperty(typeof(TProjection));
-            var fields = propertyInfos.Select(p => (p == identityProperty) ? Constants.Documents.Indexing.Fields.DocumentIdFieldName : p.Name).ToArray();
+            var fields = propertyInfos.Select(p => p == identityProperty ? Constants.Documents.Indexing.Fields.DocumentIdFieldName : p.Name).ToArray();
             return SelectFields<TProjection>(fields, projections);
         }
 
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Distinct()
         {
             Distinct();
             return this;
         }
 
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByScore()
         {
             OrderByScore();
             return this;
         }
 
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByScoreDescending()
         {
             OrderByScoreDescending();
             return this;
         }
 
+        /// <inheritdoc />
         public IDocumentQuery<T> ExplainScores()
         {
             ShouldExplainScores = true;
             return this;
         }
 
-        /// <summary>
-        /// Selects the specified fields directly from the index
-        /// </summary>
-        /// <typeparam name="TProjection">The type of the projection.</typeparam>
+        /// <inheritdoc />
         public IDocumentQuery<TProjection> SelectFields<TProjection>(params string[] fields)
         {
             return SelectFields<TProjection>(fields, fields);
         }
 
-        /// <summary>
-        /// Selects the specified fields directly from the index
-        /// </summary>
-        /// <typeparam name="TProjection">The type of the projection.</typeparam>
-        public virtual IDocumentQuery<TProjection> SelectFields<TProjection>(string[] fields, string[] projections)
+        /// <inheritdoc />
+        public IDocumentQuery<TProjection> SelectFields<TProjection>(string[] fields, string[] projections)
         {
             return CreateDocumentQueryInternal<TProjection>(fields.Length > 0 ? FieldsToFetchToken.Create(fields, projections) : null);
         }
 
-        /// <summary>
-        /// EXPERT ONLY: Instructs the query to wait for non stale results for the specified wait timeout.
-        /// This shouldn't be used outside of unit tests unless you are well aware of the implications
-        /// </summary>
-        /// <param name="waitTimeout">The wait timeout.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WaitForNonStaleResults(TimeSpan waitTimeout)
         {
             WaitForNonStaleResults(waitTimeout);
             return this;
         }
 
-        /// <summary>
-        /// Adds an ordering for a specific field to the query
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="descending">if set to <c>true</c> [descending].</param>
-        /// <param name="ordering">Ordering type.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.AddOrder(string fieldName, bool descending, OrderingType ordering)
         {
             if (descending)
@@ -111,12 +92,7 @@ namespace Raven.Client.Documents.Session
             return this;
         }
 
-        /// <summary>
-        ///   Adds an ordering for a specific field to the query
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "descending">if set to <c>true</c> [descending].</param>
-        /// <param name="ordering">Ordering type.</param>
+        /// <inheritdoc />
         public IDocumentQuery<T> AddOrder<TValue>(Expression<Func<T, TValue>> propertySelector, bool descending, OrderingType ordering)
         {
             var fieldName = GetMemberQueryPath(propertySelector.Body);
@@ -128,148 +104,119 @@ namespace Raven.Client.Documents.Session
             return this;
         }
 
-        /// <summary>
-        /// Simplified method for opening a new clause within the query
-        /// </summary>
-        /// <returns></returns>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OpenSubclause()
         {
             OpenSubclause();
             return this;
         }
 
-        /// <summary>
-        /// Simplified method for closing a clause within the query
-        /// </summary>
-        /// <returns></returns>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.CloseSubclause()
         {
             CloseSubclause();
             return this;
         }
 
-        /// <summary>
-        /// Perform a search for documents which fields that match the searchTerms.
-        /// If there is more than a single term, each of them will be checked independently.
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Search(string fieldName, string searchTerms, SearchOperator @operator)
         {
             Search(fieldName, searchTerms, @operator);
             return this;
         }
 
-        /// <summary>
-        /// Perform a search for documents which fields that match the searchTerms.
-        /// If there is more than a single term, each of them will be checked independently.
-        /// </summary>
+        /// <inheritdoc />
         public IDocumentQuery<T> Search<TValue>(Expression<Func<T, TValue>> propertySelector, string searchTerms, SearchOperator @operator)
         {
             Search(GetMemberQueryPath(propertySelector.Body), searchTerms, @operator);
             return this;
         }
 
-        /// <summary>
-        /// Partition the query so we can intersect different parts of the query
-        /// across different index entries.
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Intersect()
         {
             Intersect();
             return this;
         }
 
-        /// <summary>
-        /// Performs a query matching ANY of the provided values against the given field (OR)
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.ContainsAny(string fieldName, IEnumerable<object> values)
         {
             ContainsAny(fieldName, values);
             return this;
         }
 
-        /// <summary>
-        /// Performs a query matching ANY of the provided values against the given field (OR)
-        /// </summary>
+        /// <inheritdoc />
         public IDocumentQuery<T> ContainsAny<TValue>(Expression<Func<T, TValue>> propertySelector, IEnumerable<TValue> values)
         {
             ContainsAny(GetMemberQueryPath(propertySelector.Body), values.Cast<object>());
             return this;
         }
 
-        /// <summary>
-        /// Performs a query matching ALL of the provided values against the given field (AND)
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.ContainsAll(string fieldName, IEnumerable<object> values)
         {
             ContainsAll(fieldName, values);
             return this;
         }
 
-        /// <summary>
-        /// Performs a query matching ALL of the provided values against the given field (AND)
-        /// </summary>
+        /// <inheritdoc />
         public IDocumentQuery<T> ContainsAll<TValue>(Expression<Func<T, TValue>> propertySelector, IEnumerable<TValue> values)
         {
             ContainsAll(GetMemberQueryPath(propertySelector.Body), values.Cast<object>());
             return this;
         }
 
-        /// <summary>
-        /// Provide statistics about the query, such as total count of matching records
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Statistics(out QueryStatistics stats)
         {
             Statistics(out stats);
             return this;
         }
 
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.UsingDefaultOperator(QueryOperator queryOperator)
         {
             UsingDefaultOperator(queryOperator);
             return this;
         }
 
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.NoTracking()
         {
             NoTracking();
             return this;
         }
 
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.NoCaching()
         {
             NoCaching();
             return this;
         }
 
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.ShowTimings()
         {
             ShowTimings();
             return this;
         }
 
-        /// <summary>
-        /// Includes the specified path in the query, loading the document specified in that path
-        /// </summary>
-        /// <param name="path">The path.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Include(string path)
         {
             Include(path);
             return this;
         }
 
-        /// <summary>
-        /// Includes the specified path in the query, loading the document specified in that path
-        /// </summary>
-        /// <param name="path">The path.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Include(Expression<Func<T, object>> path)
         {
             Include(path);
             return this;
         }
 
-        /// <summary>
-        /// Negate the next operation
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Not
         {
             get
@@ -279,6 +226,7 @@ namespace Raven.Client.Documents.Session
             }
         }
 
+        /// <inheritdoc />
         public QueryResult QueryResult
         {
             get
@@ -289,484 +237,286 @@ namespace Raven.Client.Documents.Session
             }
         }
 
-        /// <summary>
-        /// Takes the specified count.
-        /// </summary>
-        /// <param name="count">The count.</param>
-        /// <returns></returns>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Take(int count)
         {
             Take(count);
             return this;
         }
 
-        /// <summary>
-        /// Skips the specified count.
-        /// </summary>
-        /// <param name="count">The count.</param>
-        /// <returns></returns>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Skip(int count)
         {
             Skip(count);
             return this;
         }
 
-        /// <summary>
-        /// Filter the results from the index using the specified where clause.
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereLucene(string fieldName, string whereClause)
         {
             WhereLucene(fieldName, whereClause);
             return this;
         }
 
-        /// <summary>
-        /// 	Matches value
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereEquals(string fieldName, object value, bool exact)
         {
             WhereEquals(fieldName, value, exact);
             return this;
         }
 
-        /// <summary>
-        /// 	Matches value
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereEquals<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value, bool exact)
         {
             WhereEquals(GetMemberQueryPath(propertySelector.Body), value, exact);
             return this;
         }
 
-        /// <summary>
-        /// 	Matches value
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereEquals(WhereParams whereParams)
         {
             WhereEquals(whereParams);
             return this;
         }
 
-        /// <summary>
-        /// 	Not matches value
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereNotEquals(string fieldName, object value, bool exact)
         {
             WhereNotEquals(fieldName, value, exact);
             return this;
         }
 
-        /// <summary>
-        /// 	Not matches value
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereNotEquals<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value, bool exact)
         {
             WhereNotEquals(GetMemberQueryPath(propertySelector.Body), value, exact);
             return this;
         }
 
-        /// <summary>
-        /// 	Not matches value
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereNotEquals(WhereParams whereParams)
         {
             WhereNotEquals(whereParams);
             return this;
         }
 
-        /// <summary>
-        /// Check that the field has one of the specified value
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereIn(string fieldName, IEnumerable<object> values, bool exact)
         {
             WhereIn(fieldName, values, exact);
             return this;
         }
 
-        /// <summary>
-        /// Check that the field has one of the specified value
-        /// </summary>
+        /// <inheritdoc />
         public IDocumentQuery<T> WhereIn<TValue>(Expression<Func<T, TValue>> propertySelector, IEnumerable<TValue> values, bool exact = false)
         {
             WhereIn(GetMemberQueryPath(propertySelector.Body), values.Cast<object>(), exact);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields which starts with the specified value.
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="value">The value.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereStartsWith(string fieldName, object value)
         {
             WhereStartsWith(fieldName, value);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields which starts with the specified value.
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "value">The value.</param>
+        /// <inheritdoc />
         public IDocumentQuery<T> WhereStartsWith<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value)
         {
             WhereStartsWith(GetMemberQueryPath(propertySelector.Body), value);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields which ends with the specified value.
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="value">The value.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereEndsWith(string fieldName, object value)
         {
             WhereEndsWith(fieldName, value);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields which ends with the specified value.
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "value">The value.</param>
+        /// <inheritdoc />
         public IDocumentQuery<T> WhereEndsWith<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value)
         {
             WhereEndsWith(GetMemberQueryPath(propertySelector.Body), value);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields where the value is between the specified start and end, exclusive
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="start">The start.</param>
-        /// <param name="end">The end.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereBetween(string fieldName, object start, object end, bool exact)
         {
             WhereBetween(fieldName, start, end, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields where the value is between the specified start and end, exclusive
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "start">The start.</param>
-        /// <param name = "end">The end.</param>
+        /// <inheritdoc />
         public IDocumentQuery<T> WhereBetween<TValue>(Expression<Func<T, TValue>> propertySelector, TValue start, TValue end, bool exact = false)
         {
             WhereBetween(GetMemberQueryPath(propertySelector.Body), start, end, exact);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields where the value is greater than the specified value
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="value">The value.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereGreaterThan(string fieldName, object value, bool exact)
         {
             WhereGreaterThan(fieldName, value, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields where the value is greater than the specified value
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "value">The value.</param>
+        /// <inheritdoc />
         public IDocumentQuery<T> WhereGreaterThan<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value, bool exact = false)
         {
             WhereGreaterThan(GetMemberQueryPath(propertySelector.Body), value, exact);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields where the value is greater than or equal to the specified value
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="value">The value.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereGreaterThanOrEqual(string fieldName, object value, bool exact)
         {
             WhereGreaterThanOrEqual(fieldName, value, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields where the value is greater than or equal to the specified value
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "value">The value.</param>
+        /// <inheritdoc />
         public IDocumentQuery<T> WhereGreaterThanOrEqual<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value, bool exact = false)
         {
             WhereGreaterThanOrEqual(GetMemberQueryPath(propertySelector.Body), value, exact);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields where the value is less than the specified value
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="value">The value.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereLessThan(string fieldName, object value, bool exact)
         {
             WhereLessThan(fieldName, value, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields where the value is less than the specified value
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "value">The value.</param>
+        /// <inheritdoc />
         public IDocumentQuery<T> WhereLessThan<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value, bool exact = false)
         {
             WhereLessThan(GetMemberQueryPath(propertySelector.Body), value, exact);
             return this;
         }
 
-        /// <summary>
-        /// Matches fields where the value is less than or equal to the specified value
-        /// </summary>
-        /// <param name="fieldName">Name of the field.</param>
-        /// <param name="value">The value.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereLessThanOrEqual(string fieldName, object value, bool exact)
         {
             WhereLessThanOrEqual(fieldName, value, exact);
             return this;
         }
 
-        /// <summary>
-        ///   Matches fields where the value is less than or equal to the specified value
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
-        /// <param name = "value">The value.</param>
+        /// <inheritdoc />
         public IDocumentQuery<T> WhereLessThanOrEqual<TValue>(Expression<Func<T, TValue>> propertySelector, TValue value, bool exact = false)
         {
             WhereLessThanOrEqual(GetMemberQueryPath(propertySelector.Body), value, exact);
             return this;
         }
 
-
-        /// <summary>
-        ///     Check if the given field exists
-        /// </summary>
-        /// <param name = "propertySelector">Property selector for the field.</param>
+        /// <inheritdoc />
         public IDocumentQuery<T> WhereExists<TValue>(Expression<Func<T, TValue>> propertySelector)
         {
             WhereExists(GetMemberQueryPath(propertySelector.Body));
             return this;
         }
 
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WhereExists(string fieldName)
         {
             WhereExists(fieldName);
             return this;
         }
 
-        /// <summary>
-        /// Add an AND to the query
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.AndAlso()
         {
             AndAlso();
             return this;
         }
 
-        /// <summary>
-        /// Add an OR to the query
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrElse()
         {
             OrElse();
             return this;
         }
 
-        /// <summary>
-        /// Specifies a boost weight to the last where clause.
-        /// The higher the boost factor, the more relevant the term will be.
-        /// </summary>
-        /// <param name="boost">boosting factor where 1.0 is default, less than 1.0 is lower weight, greater than 1.0 is higher weight</param>
-        /// <returns></returns>
-        /// <remarks>
-        /// http://lucene.apache.org/java/2_4_0/queryparsersyntax.html#Boosting%20a%20Term
-        /// </remarks>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Boost(decimal boost)
         {
             Boost(boost);
             return this;
         }
 
-        /// <summary>
-        /// Specifies a fuzziness factor to the single word term in the last where clause
-        /// </summary>
-        /// <param name="fuzzy">0.0 to 1.0 where 1.0 means closer match</param>
-        /// <returns></returns>
-        /// <remarks>
-        /// http://lucene.apache.org/java/2_4_0/queryparsersyntax.html#Fuzzy%20Searches
-        /// </remarks>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Fuzzy(decimal fuzzy)
         {
             Fuzzy(fuzzy);
             return this;
         }
 
-        /// <summary>
-        /// Specifies a proximity distance for the phrase in the last where clause
-        /// </summary>
-        /// <param name="proximity">number of words within</param>
-        /// <returns></returns>
-        /// <remarks>
-        /// http://lucene.apache.org/java/2_4_0/queryparsersyntax.html#Proximity%20Searches
-        /// </remarks>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Proximity(int proximity)
         {
             Proximity(proximity);
             return this;
         }
 
-        /// <summary>
-        /// Order the search results randomly
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.RandomOrdering()
         {
             RandomOrdering();
             return this;
         }
 
-        /// <summary>
-        /// Order the search results randomly using the specified seed
-        /// this is useful if you want to have repeatable random queries
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.RandomOrdering(string seed)
         {
             RandomOrdering(seed);
             return this;
         }
 
-        /// <summary>
-        /// Order the search results randomly
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.CustomSortUsing(string typeName, bool descending)
         {
             CustomSortUsing(typeName, descending);
             return this;
         }
 
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WithinRadiusOf<TValue>(Expression<Func<T, TValue>> propertySelector, double radius, double latitude, double longitude, SpatialUnits? radiusUnits, double distanceErrorPct)
-        {
-            WithinRadiusOf(propertySelector.ToPropertyPath(), radius, latitude, longitude, radiusUnits, distanceErrorPct);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WithinRadiusOf(string fieldName, double radius, double latitude, double longitude, SpatialUnits? radiusUnits, double distanceErrorPct)
-        {
-            WithinRadiusOf(fieldName, radius, latitude, longitude, radiusUnits, distanceErrorPct);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.RelatesToShape<TValue>(Expression<Func<T, TValue>> propertySelector, string shapeWKT, SpatialRelation relation, double distanceErrorPct)
-        {
-            Spatial(propertySelector.ToPropertyPath(), shapeWKT, relation, distanceErrorPct);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.RelatesToShape(string fieldName, string shapeWKT, SpatialRelation relation, double distanceErrorPct)
-        {
-            Spatial(fieldName, shapeWKT, relation, distanceErrorPct);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistance<TValue>(Expression<Func<T, TValue>> propertySelector, double latitude, double longitude)
-        {
-            OrderByDistance(propertySelector.ToPropertyPath(), latitude, longitude);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistance(string fieldName, double latitude, double longitude)
-        {
-            OrderByDistance(fieldName, latitude, longitude);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistance<TValue>(Expression<Func<T, TValue>> propertySelector, string shapeWkt)
-        {
-            OrderByDistance(propertySelector.ToPropertyPath(), shapeWkt);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistance(string fieldName, string shapeWkt)
-        {
-            OrderByDistance(fieldName, shapeWkt);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistanceDescending<TValue>(Expression<Func<T, TValue>> propertySelector, double latitude, double longitude)
-        {
-            OrderByDistanceDescending(propertySelector.ToPropertyPath(), latitude, longitude);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistanceDescending(string fieldName, double latitude, double longitude)
-        {
-            OrderByDistanceDescending(fieldName, latitude, longitude);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistanceDescending<TValue>(Expression<Func<T, TValue>> propertySelector, string shapeWkt)
-        {
-            OrderByDistanceDescending(propertySelector.ToPropertyPath(), shapeWkt);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDistanceDescending(string fieldName, string shapeWkt)
-        {
-            OrderByDistanceDescending(fieldName, shapeWkt);
-            return this;
-        }
-
+        /// <inheritdoc />
         IGroupByDocumentQuery<T> IDocumentQuery<T>.GroupBy(string fieldName, params string[] fieldNames)
         {
             GroupBy(fieldName, fieldNames);
             return new GroupByDocumentQuery<T>(this);
         }
 
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQuery<T>.RawQuery(string query)
         {
             RawQuery(query);
             return this;
         }
 
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQuery<T>.AddParameter(string name, object value)
         {
             AddParameter(name, value);
             return this;
         }
 
+        /// <inheritdoc />
         public IDocumentQuery<TResult> OfType<TResult>()
         {
             return CreateDocumentQueryInternal<TResult>();
         }
 
-        /// <summary>
-        /// Order the results by the specified fields
-        /// The fields are the names of the fields to sort, defaulting to sorting by ascending.
-        /// You can prefix a field name with '-' to indicate sorting by descending or '+' to sort by ascending
-        /// </summary>
-        /// <param name="fields">The fields.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderBy(string field, OrderingType ordering)
         {
             OrderBy(field, ordering);
             return this;
         }
 
-        /// <summary>
-        ///   Order the results by the specified fields
-        ///   The fields are the names of the fields to sort, defaulting to sorting by ascending.
-        ///   You can prefix a field name with '-' to indicate sorting by descending or '+' to sort by ascending
-        /// </summary>
-        /// <param name = "propertySelectors">Property selectors for the fields.</param>
+        /// <inheritdoc />
         public IDocumentQuery<T> OrderBy<TValue>(params Expression<Func<T, TValue>>[] propertySelectors)
         {
             foreach (var item in propertySelectors)
@@ -776,24 +526,14 @@ namespace Raven.Client.Documents.Session
             return this;
         }
 
-        /// <summary>
-        /// Order the results by the specified fields
-        /// The fields are the names of the fields to sort, defaulting to sorting by descending.
-        /// You can prefix a field name with '-' to indicate sorting by descending or '+' to sort by ascending
-        /// </summary>
-        /// <param name="fields">The fields.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.OrderByDescending(string field, OrderingType ordering)
         {
             OrderByDescending(field, ordering);
             return this;
         }
 
-        /// <summary>
-        ///   Order the results by the specified fields
-        ///   The fields are the names of the fields to sort, defaulting to sorting by descending.
-        ///   You can prefix a field name with '-' to indicate sorting by descending or '+' to sort by ascending
-        /// </summary>
-        /// <param name = "propertySelectors">Property selectors for the fields.</param>
+        /// <inheritdoc />
         public IDocumentQuery<T> OrderByDescending<TValue>(params Expression<Func<T, TValue>>[] propertySelectors)
         {
             foreach (var item in propertySelectors)
@@ -804,193 +544,80 @@ namespace Raven.Client.Documents.Session
             return this;
         }
 
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Highlight(
-            string fieldName,
-            int fragmentLength,
-            int fragmentCount,
-            string fragmentsField)
-        {
-            Highlight(fieldName, fragmentLength, fragmentCount, fragmentsField);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Highlight(
-            string fieldName,
-            int fragmentLength,
-            int fragmentCount,
-            out FieldHighlightings highlightings)
-        {
-            Highlight(fieldName, fragmentLength, fragmentCount, out highlightings);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Highlight(
-            string fieldName,
-            string fieldKeyName,
-            int fragmentLength,
-            int fragmentCount,
-            out FieldHighlightings highlightings)
-        {
-            Highlight(fieldName, fieldKeyName, fragmentLength, fragmentCount, out highlightings);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Highlight<TValue>(
-            Expression<Func<T, TValue>> propertySelector,
-            int fragmentLength,
-            int fragmentCount,
-            Expression<Func<T, IEnumerable>> fragmentsPropertySelector)
-        {
-            var fieldName = GetMemberQueryPath(propertySelector);
-            var fragmentsField = GetMemberQueryPath(fragmentsPropertySelector);
-            Highlight(fieldName, fragmentLength, fragmentCount, fragmentsField);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Highlight<TValue>(
-            Expression<Func<T, TValue>> propertySelector,
-            int fragmentLength,
-            int fragmentCount,
-            out FieldHighlightings fieldHighlightings)
-        {
-            Highlight(GetMemberQueryPath(propertySelector), fragmentLength, fragmentCount, out fieldHighlightings);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.Highlight<TValue>(
-            Expression<Func<T, TValue>> propertySelector,
-            Expression<Func<T, TValue>> keyPropertySelector,
-            int fragmentLength,
-            int fragmentCount,
-            out FieldHighlightings fieldHighlightings)
-        {
-            Highlight(GetMemberQueryPath(propertySelector), GetMemberQueryPath(keyPropertySelector), fragmentLength, fragmentCount, out fieldHighlightings);
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.SetHighlighterTags(string preTag, string postTag)
-        {
-            SetHighlighterTags(new[] { preTag }, new[] { postTag });
-            return this;
-        }
-
-        IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.SetHighlighterTags(string[] preTags, string[] postTags)
-        {
-            SetHighlighterTags(preTags, postTags);
-            return this;
-        }
-
-        /// <summary>
-        /// Instructs the query to wait for non stale results as of now.
-        /// </summary>
-        /// <returns></returns>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WaitForNonStaleResultsAsOfNow()
         {
             WaitForNonStaleResultsAsOfNow();
             return this;
         }
 
-        /// <summary>
-        /// Instructs the query to wait for non stale results as of now for the specified timeout.
-        /// </summary>
-        /// <param name="waitTimeout">The wait timeout.</param>
-        /// <returns></returns>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WaitForNonStaleResultsAsOfNow(TimeSpan waitTimeout)
         {
             WaitForNonStaleResultsAsOfNow(waitTimeout);
             return this;
         }
 
-        /// <summary>
-        /// Instructs the query to wait for non stale results as of the cutoff etag.
-        /// </summary>
-        /// <param name="cutOffEtag">The cut off etag.</param>
-        /// <returns></returns>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WaitForNonStaleResultsAsOf(long cutOffEtag)
         {
             WaitForNonStaleResultsAsOf(cutOffEtag);
             return this;
         }
 
-        /// <summary>
-        /// Instructs the query to wait for non stale results as of the cutoff etag for the specified timeout.
-        /// </summary>
-        /// <param name="cutOffEtag">The cut off etag.</param>
-        /// <param name="waitTimeout">The wait timeout.</param>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WaitForNonStaleResultsAsOf(long cutOffEtag, TimeSpan waitTimeout)
         {
             WaitForNonStaleResultsAsOf(cutOffEtag, waitTimeout);
             return this;
         }
 
-        /// <summary>
-        /// EXPERT ONLY: Instructs the query to wait for non stale results.
-        /// This shouldn't be used outside of unit tests unless you are well aware of the implications
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.WaitForNonStaleResults()
         {
             WaitForNonStaleResults();
             return this;
         }
 
-
-        /// <summary>
-        /// Allows you to modify the index query before it is sent to the server
-        /// </summary>
+        /// <inheritdoc />
         IDocumentQuery<T> IDocumentQueryBase<T, IDocumentQuery<T>>.BeforeQueryExecution(Action<IndexQuery> beforeQueryExecution)
         {
             BeforeQueryExecution(beforeQueryExecution);
             return this;
         }
 
-        /// <summary>
-        /// Returns an enumerator that iterates through a collection.
-        /// </summary>
-        /// <returns>
-        /// An <see cref="T:System.Collections.IEnumerator"/> object that can be used to iterate through the collection.
-        /// </returns>
-        /// <filterpriority>2</filterpriority>
+        /// <inheritdoc />
         IEnumerator IEnumerable.GetEnumerator()
         {
             return GetEnumerator();
         }
 
-        /// <summary>
-        ///   Gets the enumerator.
-        /// </summary>
-        public virtual IEnumerator<T> GetEnumerator()
+        /// <inheritdoc />
+        public IEnumerator<T> GetEnumerator()
         {
             InitSync();
             return QueryOperation.Complete<T>().GetEnumerator();
         }
 
-        public IDocumentQuery<T> Spatial(Expression<Func<T, object>> path, Func<SpatialCriteriaFactory, SpatialCriteria> clause)
-        {
-            return Spatial(path.ToPropertyPath(), clause);
-        }
-
-        public IDocumentQuery<T> Spatial(string fieldName, Func<SpatialCriteriaFactory, SpatialCriteria> clause)
-        {
-            var criteria = clause(SpatialCriteriaFactory.Instance);
-            Spatial(fieldName, criteria);
-            return this;
-        }
-
+        /// <inheritdoc />
         public T First()
         {
             return ExecuteQueryOperation(1).First();
         }
 
+        /// <inheritdoc />
         public T FirstOrDefault()
         {
             return ExecuteQueryOperation(1).FirstOrDefault();
         }
 
+        /// <inheritdoc />
         public T Single()
         {
             return ExecuteQueryOperation(2).Single();
         }
 
+        /// <inheritdoc />
         public T SingleOrDefault()
         {
             return ExecuteQueryOperation(2).SingleOrDefault();
@@ -1006,49 +633,7 @@ namespace Raven.Client.Documents.Session
             return QueryOperation.Complete<T>();
         }
 
-        public FacetedQueryResult GetFacets(string facetSetupDoc, int facetStart, int? facetPageSize)
-        {
-            var q = GetIndexQuery();
-            var query = FacetQuery.Create(q, facetSetupDoc, null, facetStart, facetPageSize, Conventions);
-
-            var command = new GetFacetsCommand(Conventions, TheSession.Context, query);
-            TheSession.RequestExecutor.Execute(command, TheSession.Context);
-
-            return command.Result;
-        }
-
-        public FacetedQueryResult GetFacets(List<Facet> facets, int facetStart, int? facetPageSize)
-        {
-            var q = GetIndexQuery();
-            var query = FacetQuery.Create(q, null, facets, facetStart, facetPageSize, Conventions);
-
-            var command = new GetFacetsCommand(Conventions, TheSession.Context, query);
-            TheSession.RequestExecutor.Execute(command, TheSession.Context);
-
-            return command.Result;
-        }
-
-        public Lazy<FacetedQueryResult> GetFacetsLazy(string facetSetupDoc, int facetStart, int? facetPageSize)
-        {
-            var q = GetIndexQuery();
-            var query = FacetQuery.Create(q, facetSetupDoc, null, facetStart, facetPageSize, Conventions);
-
-            var lazyFacetsOperation = new LazyFacetsOperation(Conventions, query);
-            return ((DocumentSession)TheSession).AddLazyOperation(lazyFacetsOperation, (Action<FacetedQueryResult>)null);
-        }
-
-        public Lazy<FacetedQueryResult> GetFacetsLazy(List<Facet> facets, int facetStart, int? facetPageSize)
-        {
-            var q = GetIndexQuery();
-            var query = FacetQuery.Create(q, null, facets, facetStart, facetPageSize, Conventions);
-
-            var lazyFacetsOperation = new LazyFacetsOperation(Conventions, query);
-            return ((DocumentSession)TheSession).AddLazyOperation(lazyFacetsOperation, (Action<FacetedQueryResult>)null);
-        }
-
-        /// <summary>
-        /// Gets the total count of records for this query
-        /// </summary>
+        /// <inheritdoc />
         public int Count()
         {
             Take(0);
@@ -1056,20 +641,14 @@ namespace Raven.Client.Documents.Session
             return queryResult.TotalResults;
         }
 
-        /// <summary>
-        /// Register the query as a lazy query in the session and return a lazy
-        /// instance that will evaluate the query only when needed
-        /// </summary>
+        /// <inheritdoc />
         public Lazy<IEnumerable<T>> Lazily()
         {
             return Lazily(null);
         }
 
-        /// <summary>
-        /// Register the query as a lazy-count query in the session and return a lazy
-        /// instance that will evaluate the query only when needed
-        /// </summary>
-        public virtual Lazy<int> CountLazily()
+        /// <inheritdoc />
+        public Lazy<int> CountLazily()
         {
             if (QueryOperation == null)
             {
@@ -1083,11 +662,8 @@ namespace Raven.Client.Documents.Session
             return ((DocumentSession)TheSession).AddLazyCountOperation(lazyQueryOperation);
         }
 
-        /// <summary>
-        /// Register the query as a lazy query in the session and return a lazy
-        /// instance that will evaluate the query only when needed
-        /// </summary>
-        public virtual Lazy<IEnumerable<T>> Lazily(Action<IEnumerable<T>> onEval)
+        /// <inheritdoc />
+        public Lazy<IEnumerable<T>> Lazily(Action<IEnumerable<T>> onEval)
         {
             if (QueryOperation == null)
             {
@@ -1098,7 +674,7 @@ namespace Raven.Client.Documents.Session
             return ((DocumentSession)TheSession).AddLazyOperation(lazyQueryOperation, onEval);
         }
 
-        protected virtual void InitSync()
+        protected void InitSync()
         {
             if (QueryOperation != null)
                 return;
@@ -1149,7 +725,6 @@ namespace Raven.Client.Documents.Session
                 TheWaitForNonStaleResults = TheWaitForNonStaleResults,
                 TheWaitForNonStaleResultsAsOfNow = TheWaitForNonStaleResultsAsOfNow,
                 Negate = Negate,
-                TransformResultsFunc = TransformResultsFunc,
                 Includes = new HashSet<string>(Includes),
                 RootTypes = { typeof(T) },
                 BeforeQueryExecutionAction = BeforeQueryExecutionAction,
