+++ test/FastTests/Client/Indexing/DebugIndexing.cs
@@ -52,10 +52,6 @@ namespace FastTests.Client.Indexing
                         CutoffEtag = q.CutoffEtag,
                         DisableCaching = q.DisableCaching,
                         ExplainScores = q.ExplainScores,
-                        HighlightedFields = q.HighlightedFields,
-                        HighlighterKeyName = q.HighlighterKeyName,
-                        HighlighterPostTags = q.HighlighterPostTags,
-                        HighlighterPreTags = q.HighlighterPreTags,
                         PageSize = q.PageSize,
                         ShowTimings = q.ShowTimings,
                         SkipDuplicateChecking = q.SkipDuplicateChecking,
